convert gdax to typescript
